Added Board.copy()
Board.copy() has been added to Board. It should make a propper deep copy of the board. The only thing not getting deep copied is the controller, but I doubt that will matter, since the controller is just there to check who owns this Land. The controllers will not change, so a controller can own Land on multiple Boards without any issues. I believe.